fix(qiniu): replace qiniu SDK with Rest API
to avoid https://github.com/qiniu/nodejs-sdk.v6/issues/205
